fix(includes): iterate airports by rule, then by airport
Currently, we iterate each of the airport directories and within that, iterate through the rules.
This has the effect of making the ordering of the rules essentially useless, as all the rules get
applied to the current airport before moving on. This change modifies the processing order so it
applies each rule in turn to each folder, which produces the desired effect.

fix #167